refs #21 -- moving things around
