+++ build_html.py
@@ -3,10 +3,10 @@ The Black Hack OGL content Python builder.
 
 This script is Public Domain.
 """
-from toolbox import Builder
+from toolbox.html import HTMLBuilder
 
 
 if __name__ == '__main__':
 
-    builder = Builder()
+    builder = HTMLBuilder()
     builder.build()
