+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = build,test
+envlist = html,test
 skipsdist = True
 
 [testenv]
@@ -9,7 +9,7 @@ deps =
     # For debuggin purposes
     # ipdb
 commands =
-    python build.py
+    python build_html.py
 
 
 [testenv:test]
